+++ packages/gator-permissions-controller/CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Refresh gator permissions map after revocation state change ([#7235](https://github.com/MetaMask/core/pull/7235))
 - Add `submitDirectRevocation` method for already-disabled delegations that don't require an on-chain transaction ([#7244](https://github.com/MetaMask/core/pull/7244))
 
 ### Changed